% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-utils.R
\name{add_website_needs}
\alias{add_website_needs}
\title{Add website Needs field to DESCRIPTION}
\usage{
add_website_needs()
}
\description{
Adds the hubverve hubStyle package as a website need to the DESCRIPTION file.
}
