% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_hubdev_readme_footer.R
\name{append_hubdev_readme_footer}
\alias{append_hubdev_readme_footer}
\title{Append the standard hubverse footer to a README.Rmd}
\usage{
append_hubdev_readme_footer(
  hubdocs_contribute_url =
    "https://hubdocs.readthedocs.io/en/latest/overview/contribute.html"
)
}
\arguments{
\item{hubdocs_contribute_url}{Character string. URL of the general contributing
information page on hubDocs.}
}
\description{
Append the standard hubverse footer to a README.Rmd
}
