% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hubdev_pkg.R
\name{create_hubdev_pkg}
\alias{create_hubdev_pkg}
\title{Create hubverse package skeleton}
\usage{
create_hubdev_pkg(
  path,
  fields = list(),
  copyright_holder = "Consortium of Infectious Disease Modeling Hubs",
  organisation = "hubverse-org",
  hubdocs_contribute_url =
    "https://hubverse.io/en/latest/overview/contribute.html"
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options.}

\item{copyright_holder}{Name of the copyright holder or holders. This
defaults to \code{"{package name} authors"}; you should only change this if you
use a CLA to assign copyright to a single entity.}

\item{organisation}{Character string. The hubverse GitHub organisation name.
The user's role and the token's scopes must be such that you have permission
to create repositories in the hubverse organisation.}

\item{hubdocs_contribute_url}{Character string. URL of the general contributing
information page on hubDocs.}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
Create hubverse package skeleton
}
