% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{use_hubdev_community}
\alias{use_hubdev_community}
\alias{use_hubdev_coc}
\alias{use_hubdev_contributing}
\title{Helpers for creating community facing hubverse package documents}
\usage{
use_hubdev_community(
  organisation = "hubverse-org",
  hubdocs_contribute_url = "https://hubverse.io/en/latest/overview/contribute.html"
)

use_hubdev_coc()

use_hubdev_contributing(
  organisation = "hubverse-org",
  hubdocs_contribute_url = "https://hubverse.io/en/latest/overview/contribute.html"
)
}
\arguments{
\item{organisation}{Character string. The hubverse GitHub organisation name.
The user's role and the token's scopes must be such that you have permission
to create repositories in the hubverse organisation.}

\item{hubdocs_contribute_url}{Character string. URL of the general contributing
information page on hubDocs.}
}
\description{
These functions create community facing documents detailing how contributors
and the wider community can interact with hubverse packages.
Each function creates the relevant file in the \code{.github} directory from templates
and is set up with appropriate hubverse default values.
}
\section{Functions}{
\itemize{
\item \code{use_hubdev_community()}: Create hubverse \code{CODE_OF_CONDUCT.md} and \code{CONTRIBUTING.md} files.

\item \code{use_hubdev_coc()}: Create hubverse \code{CODE_OF_CONDUCT.md} file.

\item \code{use_hubdev_contributing()}: Create hubverse \code{CONTRIBUTING.md} file.

}}
