% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hubdev_github.R
\name{use_hubdev_github}
\alias{use_hubdev_github}
\title{Set up hubverse package on GitHub}
\usage{
use_hubdev_github(organisation = "Infectious-Disease-Modeling-Hubs")
}
\arguments{
\item{organisation}{Character string. The hubverse GitHub organisation name.
The user's role and the token's scopes must be such that you have permission
to create repositories in the hubverse organisation.}
}
\description{
The function runs a number of internal utilities for setting up a hubverse package on GitHub:
\itemize{
\item Creates a repo in the hubverse GitHub organisation
\item Adds details of the repository to the DESCRIPTION file
\item Initailises pkgdown documentation
\item Creates GitHub Action workflows for:
\itemize{
\item standard R CMD CHECK
\item test coverage
\item linting with \code{lintr}
\item building pkgdown documentation and deploying production docs to GitHub Pages
and PR previews to a Netlify site
}
}
}
