% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hubdev_github_action.R
\name{use_hubdev_github_action}
\alias{use_hubdev_github_action}
\title{Hubverse Developer GitHub Action setup}
\usage{
use_hubdev_github_action(name)
}
\arguments{
\item{name}{Name of workflow, i.e. the name of one of the
\href{https://github.com/Infectious-Disease-Modeling-Hubs/hubverse-developer-actions}{hubverse developer action repository} directories containing a GitHub Action workflow \code{.yaml} file.}
}
\description{
Sets up common hubverse developer continuous integration (CI) workflows for
hubverse packages hosted on GitHub using
\href{https://github.com/features/actions}{GitHub Actions}.
Available actions are hosted in repository \href{https://github.com/Infectious-Disease-Modeling-Hubs/hubverse-developer-actions}{Infectious-Disease-Modeling-Hubs/hubverse-developer-actions}
}
