% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hubdev_pkgdown_action.R
\name{use_hubdev_pkgdown_action}
\alias{use_hubdev_pkgdown_action}
\title{Use the hubdev custom pkgdown GitHub Action}
\usage{
use_hubdev_pkgdown_action(remove_std = TRUE)
}
\arguments{
\item{remove_std}{Logical. Whether to remove the standard pkgdown action if it exists.}
}
\description{
The functions creates a \code{pkgdown-netlify-preview.yaml} action in the \code{.github/workflows}
directory. The action follows the standard pkgdown publishing to \code{gh-pages} in the
event of a merge into \code{main} branch but can also be set up to publish internal
PR previews to Netlify.
}
\details{
To activate Netlify Previews, you must:
\itemize{
\item Create a new site on Netlify. See \url{https://docs.netlify.com/welcome/add-new-site/} for more details.
\item "Add the netlify API site ID to the GitHub repository Action secrets as
\code{NETLIFY_SITE_ID}. See \url{https://docs.netlify.com/api/get-started/#get-site} for more details.
\item Get developer token from Netlify developer account settings and add it to
GitHub repository Action secrets as \code{NETLIFY_AUTH_TOKEN}. See
\url{https://docs.netlify.com/cli/get-started/#obtain-a-token-in-the-netlify-ui}
for more details.
}

See  \url{pkgdown} docs section on
\href{https://pkgdown.r-lib.org/articles/customise.html#pr-previews}{PR previews}
for more details.
}
