% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hubdev_readme.R
\name{use_hubdev_readme}
\alias{use_hubdev_readme}
\title{Create a basic hubverse README.Rmd from template}
\usage{
use_hubdev_readme(
  organisation = "hubverse-org",
  hubdocs_contribute_url = "https://hubverse.io/community/",
  build = TRUE
)
}
\arguments{
\item{organisation}{Character string. The hubverse GitHub organisation name.
The user's role and the token's scopes must be such that you have permission
to create repositories in the hubverse organisation.}

\item{hubdocs_contribute_url}{Character string. URL of the general contributing
information page on hubDocs.}

\item{build}{Logical. Whether to build the README.md from README.Rmd.
Defaults to \code{TRUE}.}
}
\description{
Create a basic hubverse README.Rmd from template
}
