% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hubdev_pkgdown.R
\name{use_hubdev_pkgdown}
\alias{use_hubdev_pkgdown}
\title{Add hubverse _pkgdown.yml}
\usage{
use_hubdev_pkgdown()
}
\description{
Adds a customised _pkgdown.yml file to the package root directory configured to
use the hubverse hubStyle package for docs styling.
}
