% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_hubdev_readme.R
\name{use_hubdev_readme}
\alias{use_hubdev_readme}
\title{Create a basic hubverse README.Rmd from template}
\usage{
use_hubdev_readme(
  organisation = "Infectious-Disease-Modeling-Hubs",
  hubdocs_contribute_url =
    "https://hubdocs.readthedocs.io/en/latest/overview/contribute.html"
)
}
\arguments{
\item{organisation}{Character string. The hubverse GitHub organisation name.
The user's role and the token's scopes must be such that you have permission
to create repositories in the hubverse organisation.}

\item{hubdocs_contribute_url}{Character string. URL of the general contributing
information page on hubDocs.}
}
\description{
Create a basic hubverse README.Rmd from template
}
